+++ packages/core/src/decorators/param.ts
@@ -1,7 +1,30 @@
 import { TargetParamFunction, isString } from '@koa-ioc/misc'
 import { Decorator, ParamPath } from '../constants'
-import { ParamMetadata, PipeOptions } from '../type'
+import { ParamHandle, ParamMetadata, PipeOptions } from '../type'
 import { Pipe } from './pipe'
+
+export function defineParamDecorator<T>(handle: ParamHandle<T>) {
+  return function (data?: T): TargetParamFunction {
+    return function (target, methodName, index) {
+      const paramMetadata: ParamMetadata =
+        Reflect.getMetadata(Decorator.Param, target.constructor, methodName) ||
+        []
+
+      paramMetadata.push({
+        handle,
+        index,
+        data,
+      })
+
+      Reflect.defineMetadata(
+        Decorator.Param,
+        paramMetadata,
+        target.constructor,
+        methodName
+      )
+    }
+  }
+}
 
 function createParamDecorator<T extends string>(paramPath: string) {
   function ParamDecorator(): TargetParamFunction
@@ -12,30 +35,17 @@ function createParamDecorator<T extends string>(paramPath: string) {
     ...pipes: PipeOptions[]
   ): TargetParamFunction {
     return function (target, methodName, index) {
-      const paramMetadata: ParamMetadata =
-        Reflect.getMetadata(Decorator.Param, target.constructor, methodName) ||
-        []
       if (isString(name)) {
-        paramMetadata.push({
-          paramPath,
-          index,
-          name,
-        })
         Pipe(...pipes)(target, methodName, index)
       } else {
-        paramMetadata.push({
-          paramPath,
-          index,
-        })
         Pipe(...(name ? [name] : []), ...pipes)(target, methodName, index)
       }
+      defineParamDecorator((_, ctx) => {
+        const paths = paramPath.split('.')
+        const value = paths.reduce<any>((prev, nextPath) => prev[nextPath], ctx)
 
-      Reflect.defineMetadata(
-        Decorator.Param,
-        paramMetadata,
-        target.constructor,
-        methodName
-      )
+        return isString(name) ? value?.[name] : value
+      })()(target, methodName, index)
     }
   }
   return ParamDecorator